requirements.txt
python3-tk
socket
threading
pillow
subprocess

